V0.01.1

+++ Pathways.cpp
@@ -11,7 +11,7 @@
 std::vector<std::pair<std::string,std::string>> Pathways::pathways;
 
 std::string Pathways::get(std::string key) {
-    for (int i = 0; i < pathways.size(); i++){
+    for (size_t i = 0; i < pathways.size(); i++){
         if (pathways[i].first == key)
             return pathways[i].second;
     }
